Script - Fix decimal addition calculation

+++ script.js
@@ -62,7 +62,7 @@ const calculate = () => {
   let result = "";
   switch (calculationOperator) {
     case "+":
-      result = parseInt(prevNumber) + parseInt(currentNumber);
+      result = parseFloat(prevNumber) + parseFloat(currentNumber);
       break;
     case "-":
       result = prevNumber - currentNumber;
